refactor: reorganized game logic for better maintainability
