+++ game/client/src/contexts/DojoContext.tsx
@@ -5,9 +5,9 @@ import {
     useBurnerManager,
 } from "@dojoengine/create-burner";
 import { Account } from "starknet";
-import { dojoConfig } from "../dojoConfig";
+import { dojoConfig } from "../../dojoConfig";
 import { DojoProvider } from "@dojoengine/core";
-import { setupWorld } from "./typescript/contracts.gen";
+import { setupWorld } from "../typescript/contracts.gen";
 
 /**
  * Interface defining the shape of the Dojo context.
