+++ game/client/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/dojocontext.tsx","./src/historical-events.tsx","./src/main.tsx","./src/starknet-provider.tsx","./src/usesystemcalls.ts","./src/vite-env.d.ts","./src/wallet-account.tsx","./src/typescript/contracts.gen.ts","./src/typescript/models.gen.ts"],"version":"5.7.2"}
+{"root":["./src/router.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/historicalevents.tsx","./src/components/walletaccount.tsx","./src/components/game/gamecontrols.tsx","./src/components/game/gameinfo.tsx","./src/components/game/gamestats.tsx","./src/components/home/hero.tsx","./src/components/home/howtoplay.tsx","./src/components/home/leaderboard.tsx","./src/components/home/socialmedia.tsx","./src/components/shared/footer.tsx","./src/components/shared/header.tsx","./src/contexts/dojocontext.tsx","./src/contexts/starknetprovider.tsx","./src/hooks/usegamedata.ts","./src/hooks/usesystemcalls.ts","./src/pages/game.tsx","./src/pages/home.tsx","./src/typescript/contracts.gen.ts","./src/typescript/models.gen.ts"],"version":"5.7.3"}
