+++ game/client/tsconfig.node.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./vite.config.ts"],"version":"5.7.2"}
+{"root":["./vite.config.ts"],"version":"5.7.3"}
